Rename kafka_files_handler => kafka_files_generator
