+++ protocol/kafka_files_generator/cluster_metadata_generator.go
@@ -1,4 +1,4 @@
-package kafka_files_handler
+package kafka_files_generator
 
 import (
 	"fmt"
