+++ protocol/kafka_files_generator/files_handler.go
@@ -1,4 +1,4 @@
-package kafka_files_handler
+package kafka_files_generator
 
 import (
 	"fmt"
