+++ protocol/kafka_files_generator/log_directory_generator.go
@@ -1,4 +1,4 @@
-package kafka_files_handler
+package kafka_files_generator
 
 import "github.com/codecrafters-io/tester-utils/logger"
