+++ protocol/kafka_files_generator/partition_generator.go
@@ -1,4 +1,4 @@
-package kafka_files_handler
+package kafka_files_generator
 
 import (
 	"fmt"
